refactor: WebSocketConfig 내에 있던 Redis 설정 Redis로 이관
