+++ src/main/java/mafia/mafiatogether/config/RedisConfig.java
@@ -1,11 +1,15 @@
 package mafia.mafiatogether.config;
 
+import mafia.mafiatogether.chat.application.RedisMessageBroker;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.listener.PatternTopic;
+import org.springframework.data.redis.listener.RedisMessageListenerContainer;
+import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
 import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
 import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
@@ -43,4 +47,22 @@ public class RedisConfig {
 
         return template;
     }
+
+    @Bean
+    public MessageListenerAdapter redisMessageListenerAdapter(RedisMessageBroker redisMessageBroker) {
+        return new MessageListenerAdapter(redisMessageBroker, "onMessage");
+    }
+
+    @Bean
+    public RedisMessageListenerContainer redisMessageListenerContainer(
+            RedisConnectionFactory redisConnectionFactory,
+            MessageListenerAdapter redisMessageListenerAdapter
+    ) {
+        RedisMessageListenerContainer container = new RedisMessageListenerContainer();
+        container.setConnectionFactory(redisConnectionFactory);
+        container.addMessageListener(redisMessageListenerAdapter, new PatternTopic("/sub/**"));
+
+        return container;
+    }
+
 }
