--- src/main/java/mafia/mafiatogether/config/WebsocketConfig.java
@@ -1,58 +1,0 @@
-package mafia.mafiatogether.config;
-
-
-import lombok.RequiredArgsConstructor;
-import mafia.mafiatogether.chat.application.RedisMessageBroker;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.listener.PatternTopic;
-import org.springframework.data.redis.listener.RedisMessageListenerContainer;
-import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
-import org.springframework.messaging.simp.config.ChannelRegistration;
-import org.springframework.messaging.simp.config.MessageBrokerRegistry;
-import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
-import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
-import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
-
-@Configuration
-@EnableWebSocketMessageBroker
-@RequiredArgsConstructor
-public class WebsocketConfig implements WebSocketMessageBrokerConfigurer {
-
-    private final StompChannelInterceptor stompChannelInterceptor;
-
-    @Override
-    public void registerStompEndpoints(StompEndpointRegistry registry) {
-        registry.addEndpoint("/stomp");
-    }
-
-    @Override
-    public void configureMessageBroker(MessageBrokerRegistry registry) {
-        registry.setApplicationDestinationPrefixes("/pub");
-        registry.enableSimpleBroker("/sub");
-    }
-
-    @Override
-    public void configureClientInboundChannel(ChannelRegistration registry) {
-        registry.interceptors(stompChannelInterceptor);
-    }
-
-    @Bean
-    public MessageListenerAdapter redisMessageListenerAdapter(RedisMessageBroker redisMessageBroker) {
-        return new MessageListenerAdapter(redisMessageBroker, "onMessage");
-    }
-
-    @Bean
-    public RedisMessageListenerContainer redisMessageListenerContainer(
-            RedisConnectionFactory redisConnectionFactory,
-            MessageListenerAdapter redisMessageListenerAdapter
-    ) {
-        RedisMessageListenerContainer container = new RedisMessageListenerContainer();
-        container.setConnectionFactory(redisConnectionFactory);
-        container.addMessageListener(redisMessageListenerAdapter, new PatternTopic("/sub/**"));
-
-        return container;
-    }
-
-}
